refactor: move vectorfield
